Define the new RepositoryCommitSummaryPage data structure.
This will represent one page of results from the repository commit tree.
Of note here, it will now inform callers of the total number of commits
and echo back their maxCount and skip parameters as a part of the
response.

I opted to put these into the response payload instead of the headers to
keep the model simpler for the client code. Our client code today just
returns a deserialized JSON payload. If we start having to consider
headers things get decidedly less simple.