Commit counting (#8)
This changes the data structure we return for the commit list request so that it surfaces enough data for calling code to make smart predictions about when they will reach the end of the commit pagination. It also might be interesting for calling code to surface the total number of commits in the repo.