Finicky Finagle: Implement a Finagle HTTP driver package. (#9)
This implements support for using Finagle as the underlying Giterrific
Client's HTTP driver. In further support of our pluggable dependency
pattern we want to meet consumers of the library "where they're at."
If Finagle is their preferred HTTP implementation that they're already
pulling in, then they should be able to pull in this package in
addition to Giterrific instead of having to add Databinder Dispatch to
their list of dependencies.